requirements.txt
anyio==3.6.2
certifi==2023.5.7
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.1
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pydantic==1.10.7
pytest==7.3.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.22.0
